+++ package.json
@@ -84,7 +84,6 @@
     "@vitest/coverage-v8": "^2.0.2",
     "@vue/test-utils": "^2.4.6",
     "alpinejs": "^3.14.0",
-    "happy-dom": "^14.12.0",
     "jsdom": "^25.0.1",
     "svelte": "^4.2.17",
     "ts-standard": "^12.0.2",
